Family{C}: generalize constructor method
This patch changes how new child collectors are constructed. Previously
only `metric_name`, `help`, and `registry=nothing` were passed to the
constructor `C`. With this patch any `args...` and `; kwargs...` not
consumed by the `Family{C}` constructor are passed along as well. In
order to make this efficient a constructor closure is created in the
`Family{C}` constructor and stored in the `Family` struct.

This change is in preparation for the Histogram collector where more
arguments is needed to construct the children. It also make it easier to
later add the option to explicitly pass a constructor, or to specialize
the closure on the specific collector type `C`.