ProcessCollector: initialize on first use
This patch changes how the ProcessCollector is initialized. Previously
some values where read from /proc filesystem in the constructor but
since typically a collector is defined as a global variable this results
in values beeing cached from precompilation. With this patch the
collector is initialized on first use (first `Prometheus.collect!`)
instead.